refactor: move handlers to command handler and query handler directories
