+++ packages/core/src/common/application/commandHandler/DeleteCommandHandler.ts
