+++ packages/core/src/index.ts
@@ -1,8 +1,8 @@
-import { DeleteCommandHandler } from './common/application/handler/DeleteCommandHandler';
-import { FindOneQueryHandler } from './common/application/handler/FindOneQueryHandler';
-import { FindQueryHandler } from './common/application/handler/FindQueryHandler';
-import { InsertOneCommandHandler } from './common/application/handler/InsertOneCommandHandler';
-import { UpdateCommandHandler } from './common/application/handler/UpdateCommandHandler';
+import { DeleteCommandHandler } from './common/application/commandHandler/DeleteCommandHandler';
+import { InsertOneCommandHandler } from './common/application/commandHandler/InsertOneCommandHandler';
+import { UpdateCommandHandler } from './common/application/commandHandler/UpdateCommandHandler';
+import { FindOneQueryHandler } from './common/application/queryHandler/FindOneQueryHandler';
+import { FindQueryHandler } from './common/application/queryHandler/FindQueryHandler';
 import { DeleteAdapter } from './common/domain/adapter/DeleteAdapter';
 import { FindAdapter } from './common/domain/adapter/FindAdapter';
 import { InsertAdapter } from './common/domain/adapter/InsertAdapter';
